refactor: move money util functions to own file
