+++ src/utils/index.js
@@ -8,25 +8,5 @@ const getInitials = fullName => {
   return initials;
 };
 
-const moneyFormatter = num => {
-  if (num >= 1000000000) {
-    return `${(num / 1000000000).toFixed(1).replace(/\.0$/, '')}G`;
-  }
-  if (num >= 1000000) {
-    return `${(num / 1000000).toFixed(1).replace(/\.0$/, '')}M`;
-  }
-  if (num >= 1000) {
-    return `${(num / 1000).toFixed(1).replace(/\.0$/, '')}K`;
-  }
-  return parseFloat(num).toFixed(2);
-};
-
-const validateMoneyString = string => {
-  if (string) {
-    return string.match(/^[0-9]+(\.[0-9]{1,2})?$/gm);
-  }
-  return false;
-};
-
 // eslint-disable-next-line import/prefer-default-export
-export { getInitials, moneyFormatter, validateMoneyString };
+export { getInitials };
